--- lib/web3/settings.js
@@ -1,9 +1,0 @@
-
-
-var Settings = function () {
-    this.defaultBlock = 'latest';
-    this.defaultAccount = undefined;
-};
-
-module.exports = Settings;
-
