+++ packages/web3-core-helpers/package.json
@@ -1,10 +1,12 @@
 {
-    "name": "web3-core-helper",
+    "name": "web3-core-helpers",
     "version": "1.0.0",
     "description": "Web3 core tools helper for sub packages. This is an internal package.",
-    "repository": "https://github.com/ethereum/web3.js/tree/master/packages/web3-core-helper",
+    "repository": "https://github.com/ethereum/web3.js/tree/master/packages/web3-core-helpers",
     "license": "LGPL-3.0",
     "main": "src/index.js",
     "dependencies": {
+        "web3-utils": "^1.0.0",
+        "web3-iban": "^1.0.0"
     }
 }
