+++ packages/web3-core-helpers/src/errors.js
