+++ packages/web3-core-requestManager/src/batch.js
@@ -21,7 +21,7 @@
  */
 
 var Jsonrpc = require('./jsonrpc');
-var errors = require('web3-core').errors;
+var errors = require('web3-core-helpers').errors;
 
 var Batch = function (web3) {
     this.requestManager = web3._requestManager;
