+++ packages/web3-core/package.json
@@ -6,6 +6,9 @@
     "license": "LGPL-3.0",
     "main": "src/index.js",
     "dependencies": {
-        "web3-core-helper": "^1.0.0"
+        "web3-core-helpers": "^1.0.0",
+        "web3-core-method": "^1.0.0",
+        "web3-utils": "^1.0.0",
+        "web3-requestManager": "^1.0.0"
     }
 }
