+++ packages/web3-eth/package.json
@@ -7,9 +7,11 @@
     "main": "src/index.js",
     "dependencies": {
         "web3-core": "^1.0.0",
+        "web3-core-helpers": "^1.0.0",
         "web3-core-subscriptions": "^1.0.0",
         "web3-core-method": "^1.0.0",
         "web3-utils": "^1.0.0",
-        "lodash": "^4.17.4"
+        "lodash": "^4.17.4",
+        "eventemitter3": "^1.1.1"
     }
 }
