+++ packages/web3-personal/package.json
@@ -7,6 +7,7 @@
     "main": "src/index.js",
     "dependencies": {
         "web3-core": "^1.0.0",
+        "web3-core-helpers": "^1.0.0",
         "web3-core-method": "^1.0.0"
     }
 }
